requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.0.7
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
google-auth==2.37.0
gunicorn==22.0.0
idna==3.10
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
python-dotenv==1.0.1
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.3.0
